new database index changed

+++ install/db.sql
@@ -65,7 +65,7 @@ CREATE TABLE IF NOT EXISTS `{BO_DB_PREF}strikes` (
   `id` int(11) unsigned NOT NULL auto_increment,
   `time` datetime NOT NULL,
   `time_ns` int(11) NOT NULL,
-  `time_key` mediumint(9) unsigned NOT NULL,
+  `time_key` smallint(6) unsigned NOT NULL,
   `lat` decimal(9,6) NOT NULL,
   `lon` decimal(9,6) NOT NULL,
   `lat2` tinyint(4) NOT NULL,
